requirements.txt
asgiref==3.4.1
boto3==1.20.20
botocore==1.23.20
Django==3.2.9
djangorestframework==3.12.4
jmespath==0.10.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.7
watchtower==2.0.1
